+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GenerateDocumentationFile
 - Split files
+- Use EditorBrowsable(EditorBrowsableState.Never) instead of private or internal
 
 ## [1.0.5] - 2020-12-18
 ### Added